pyproject.toml
[project]
name = "pmtb"
version = "0.1.0"
description = "Prediction Market Trading Bot — Kalshi automated trading system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kalshi-python-async>=3.8.0",
    "websockets>=14.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic-settings[yaml]>=2.0",
    "loguru>=0.7",
    "tenacity>=8.0",
    "prometheus-client>=0.20",
    "cryptography>=42",
    "pyyaml>=6.0",
    "httpx>=0.28.1",
    "vadersentiment>=3.3.2",
    "anthropic>=0.84.0",
    "asyncpraw>=7.8.1",
    "feedparser>=6.0.12",
    "pytrends>=4.9.2",
    "xgboost>=3.2.0",
    "scikit-learn>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=1.0",
    "pytest-mock>=3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pmtb"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "demo: requires demo API credentials",
]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=1.0",
    "pytest-mock>=3.0",
]
